requirements.txt
django<1.11
python-firebase==1.2
